# api-backend/src/main/java/com/searetailer/api/model/dto/ErrorResponse.java
package com.searetailer.api.model.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.OffsetDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ErrorResponse {

    private String error;
    private String message;

    @JsonProperty("request_id")
    private String requestId;

    private OffsetDateTime timestamp;
}