# shared-contracts/openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: SEA Retailer Top Selling Intelligence API
  version: 1.0.0
  description: |
    API for the SEA Retailer Top Selling Intelligence platform.
    Provides cross-platform top-selling item analysis, matching, alerts, and review workflows.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.sea-retailer.com/api/v1
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC/JWT Bearer token
...
  schemas:
    Platform:
      type: string
      enum: [tiktok, shopee]
    MatchType:
      type: string
      enum: [exact_same, variant_family, similar, no_match]
    MatchStatus:
      type: string
      enum: [auto_accepted, needs_review, no_match, overridden]
...
    MatchReasons:
      type: object
      properties:
        strong_evidence:
          type: array
          items:
            type: string
        weak_evidence:
          type: array
          items:
            type: string
        field_alignment:
          type: object
          properties:
            brand_match:
              type: boolean
            brand_a:
              type: string
            brand_b:
              type: string
            spec_match:
              type: boolean
            spec_a:
              type: string
            spec_b:
              type: string
            title_similarity:
              type: number
            price_band_match:
              type: boolean
        missing_fields:
          type: array
          items:
            type: string
    TopItem:
      type: object
      required: [event_date, platform, item_id, title, category, rank]
      properties:
        event_date:
          type: string
          format: date
        platform:
          $ref: "#/components/schemas/Platform"
        item_id:
          type: string
        title:
          type: string
        category:
          type: string
        rank:
          type: integer
        url:
          type: string
        brand_raw:
          type: string
        brand_std:
          type: string
        size_value:
          type: number
        size_unit:
          type: string
        pack_count:
          type: integer
        image_url:
          type: string
        list_price:
          type: integer
          description: Price in smallest currency unit
        promo_price:
          type: integer
        currency:
          type: string
        review_count:
          type: integer
        sold_range:
          type: string
        source_type:
          type: string
          enum: [api, scrape, stub]
    TopItemDetail:
      allOf:
        - $ref: "#/components/schemas/TopItem"
        - type: object
          properties:
            price_history:
              type: array
              items:
                type: object
                properties:
                  event_date:
                    type: string
                    format: date
                  list_price:
                    type: integer
                  promo_price:
                    type: integer
            proxy_history:
              type: array
              items:
                type: object
                properties:
                  event_date:
                    type: string
                    format: date
                  proxy_type:
                    type: string
                  proxy_value:
                    type: string
            matches:
              type: array
              items:
                $ref: "#/components/schemas/MatchResult"
    MatchResult:
      type: object
      properties:
        tiktok_item_id:
          type: string
        shopee_item_id:
          type: string
        tiktok_title:
          type: string
        shopee_title:
          type: string
        match_type:
          $ref: "#/components/schemas/MatchType"
        confidence:
          type: number
        status:
          $ref: "#/components/schemas/MatchStatus"
        reasons:
          $ref: "#/components/schemas/MatchReasons"
        rule_version:
          type: string
...
    ReviewDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [accept, reject, change_type]
        new_match_type:
          $ref: "#/components/schemas/MatchType"
          description: Required when decision is change_type
        comment:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
paths:
...
  /items/{platform}/{item_id}:
    get:
      operationId: getItemDetail
      summary: Get item detail with history and matches
      tags: [Top Items]
      x-roles: [viewer, reviewer, admin]
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Platform"
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopItemDetail"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /review/{tiktok_item_id}/{shopee_item_id}/decision:
    post:
      operationId: submitReviewDecision
      summary: Submit a review decision for a matching pair
      tags: [Review]
      x-roles: [reviewer, admin]
      parameters:
        - name: tiktok_item_id
          in: path
          required: true
          schema:
            type: string
        - name: shopee_item_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          schema:
            type: string
          description: Unique key to prevent duplicate submissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  decision_id:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        "409":
          description: Duplicate submission (idempotency key conflict)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...